fix: should update outline panel notes after slicing note (#8327)
Fix: [BS-1358](https://linear.app/affine-design/issue/BS-1358/[bug]-使用剪刀切开之后-note-block-不会显示在-frame-toc-里) [BS-1002](https://linear.app/affine-design/issue/BS-1002/使用剪刀后，会触发选区)

Split outline.spec.ts tests into two files.
